+++ CHANGELOG.md
@@ -5,9 +5,10 @@ to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- 2d5fe52 WIP: IncreasedChecker Fix
 - 65f3836 Refactor commits
 - 0634b57 Update Increaded Checker and Defaultchecker
-- 84aaa01 WIP: rewoke the IncreasedChecker
+- 84aaa01 WIP: rewoke the ChangeChecker
 ## [1.0.0] - 2021-06-28
 ### Added
 - Added Changelog Checkers## [0.0.4] - 2021-06-27